backend/pyproject.toml
[project]
name = "whoami-backend"
version = "0.1.0"
description = "whoami - Cross-AI user identity profile sync service backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.7.0",
    "authlib>=1.4.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "cos-python-sdk-v5>=1.9.0",
    "qcloud-python-sts>=3.1.0",
    "cachetools>=5.5.0",
    "python-multipart>=0.0.18",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
